pyproject.toml
[project]
name = "misinformation-detector"
version = "0.1.0"
description = "FastAPI and React app for misinformation detection."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "python-jose",
    "passlib[bcrypt]",
    "pydantic-settings",
    "python-multipart",
    "sqlmodel",
    "uvicorn",
]

[project.optional-dependencies]
dev = [
    "httpx",
    "pytest",
    "ruff",
]

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
addopts = "-p no:cacheprovider"

[tool.ruff]
line-length = 100
target-version = "py310"
extend-exclude = ["backend/app/ml/training.ipynb"]

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = ["B008"]
